requirements.txt
Django==2.2
djangorestframework==3.12.2
djangorestframework-simplejwt==4.3.0
requests==2.25.0
